fix: update modify start time and end time of epoch in api current epoch and delegation header

+++ src/main/java/org/cardanofoundation/explorer/api/model/response/pool/DelegationHeaderResponse.java
@@ -15,7 +15,7 @@ public class DelegationHeaderResponse implements Serializable {
 
   private Long countDownEndTime;
 
-  private Long epochSlotNo;
+  private Integer epochSlotNo;
 
   private BigInteger liveStake;
